requirements.txt
Jinja2
pyyaml
pydantic
pytest
docker
click
